+++ src/crawlers/basic_crawler.js
@@ -2,12 +2,12 @@ import { ACTOR_EVENT_NAMES } from '@apify/consts';
 import { cryptoRandomObjectId } from '@apify/utilities';
 import ow, { ArgumentError } from 'ow';
 import _ from 'underscore';
+import { addTimeoutToPromise, tryCancel } from '@apify/timeout';
 import AutoscaledPool from '../autoscaling/autoscaled_pool'; // eslint-disable-line import/no-duplicates
 import events from '../events';
 import { openSessionPool } from '../session_pool/session_pool'; // eslint-disable-line import/no-duplicates
 import Statistics from './statistics';
-import { addTimeoutToPromise } from '../utils';
-import defaultLog from '../utils_log';
+import defaultLog from '../utils_log'; // eslint-disable-line import/no-duplicates
 import { validators } from '../validators';
 
 // TYPE IMPORTS
@@ -20,6 +20,7 @@ import { RequestQueue } from '../storages/request_queue';
 import { QueueOperationInfo } from '../storages/request_queue';
 import { Session } from '../session_pool/session';
 import { SessionPoolOptions } from '../session_pool/session_pool';
+import { Log } from '../utils_log';
 /* eslint-enable no-unused-vars,import/named,import/no-duplicates,import/order */
 
 /**
@@ -254,13 +255,14 @@ export class BasicCrawler {
         }
 
         // assigning {} to the options as default break proper typing
-        /** @type {defaultLog.Log} */
+        /** @type {Log} */
         this.log = log;
         this.requestList = requestList;
         this.requestQueue = requestQueue;
         this.userProvidedHandler = handleRequestFunction;
         this.failedContextHandler = handleFailedRequestFunction;
         this.handleRequestTimeoutMillis = handleRequestTimeoutSecs * 1000;
+        this.internalTimeoutMillis = Math.max(this.handleRequestTimeoutMillis, 300e3); // allow at least 5min for internal timeouts
         this.handleFailedRequestFunction = handleFailedRequestFunction;
         this.maxRequestRetries = maxRequestRetries;
         this.handledRequestsCount = 0;
@@ -289,7 +291,7 @@ export class BasicCrawler {
         const basicCrawlerAutoscaledPoolConfiguration = {
             minConcurrency,
             maxConcurrency,
-            runTaskFunction: this.timeoutForRunTaskFunction.bind(this),
+            runTaskFunction: this._runTaskFunction.bind(this),
             isTaskReadyFunction: async () => {
                 if (isMaxPagesExceeded()) {
                     if (shouldLogMaxPagesExceeded) {
@@ -460,26 +462,8 @@ export class BasicCrawler {
             await this.requestList.reclaimRequest(request);
             return null;
         }
-        const [nextRequest] = await Promise.all([
-            this.requestQueue.fetchNextRequest(),
-            this.requestList.markRequestHandled(request),
-        ]);
-        return nextRequest;
-    }
-
-    /**
-     * Wrapper around _runTaskFunction to set a timeout on the entire function.
-     * Temporary solution to see whether this causes the 0 concurrency bug.
-     * @ignore
-     * @protected
-     * @interface
-     */
-    async timeoutForRunTaskFunction() {
-        await addTimeoutToPromise(
-            this._runTaskFunction(),
-            this.handleRequestTimeoutMillis * 2,
-            `_runTaskFunction timed out after ${(this.handleRequestTimeoutMillis * 2) / 1000} seconds.`,
-        );
+        await this.requestList.markRequestHandled(request);
+        return this.requestQueue.fetchNextRequest();
     }
 
     /**
@@ -496,11 +480,27 @@ export class BasicCrawler {
         let request;
         let session;
 
+        await this._timeoutAndRetry(
+            async () => {
+                request = await this._fetchNextRequest();
+            },
+            this.internalTimeoutMillis,
+            `Fetching next request timed out after ${this.internalTimeoutMillis / 1e3} seconds.`,
+        );
+
+        tryCancel();
+
         if (this.useSessionPool) {
-            [request, session] = await Promise.all([this._fetchNextRequest(), this.sessionPool.getSession()]);
-        } else {
-            request = await this._fetchNextRequest();
+            await this._timeoutAndRetry(
+                async () => {
+                    session = await this.sessionPool.getSession();
+                },
+                this.internalTimeoutMillis,
+                `Fetching session timed out after ${this.internalTimeoutMillis / 1e3} seconds.`,
+            );
         }
+
+        tryCancel();
 
         if (!request) return;
 
@@ -521,11 +521,17 @@ export class BasicCrawler {
 
         try {
             await addTimeoutToPromise(
-                this._handleRequestFunction(crawlingContext),
+                () => this._handleRequestFunction(crawlingContext),
                 this.handleRequestTimeoutMillis,
-                `handleRequestFunction timed out after ${this.handleRequestTimeoutMillis / 1000} seconds.`,
+                `handleRequestFunction timed out after ${this.handleRequestTimeoutMillis / 1000} seconds (${request.id}).`,
             );
-            await source.markRequestHandled(request);
+
+            await this._timeoutAndRetry(
+                () => source.markRequestHandled(request),
+                this.internalTimeoutMillis,
+                `Marking request ${request.url} (${request.id}) as handled timed out after ${this.internalTimeoutMillis / 1e3} seconds.`,
+            );
+
             this.stats.finishJob(statisticsId);
             this.handledRequestsCount++;
 
@@ -533,7 +539,11 @@ export class BasicCrawler {
             if (session) session.markGood();
         } catch (err) {
             try {
-                await this._requestFunctionErrorHandler(err, crawlingContext, source);
+                await this._timeoutAndRetry(
+                    () => this._requestFunctionErrorHandler(err, crawlingContext, source),
+                    this.internalTimeoutMillis,
+                    `Handling request failure of ${request.url} (${request.id}) timed out after ${this.internalTimeoutMillis / 1e3} seconds.`,
+                );
             } catch (secondaryError) {
                 this.log.exception(secondaryError, 'runTaskFunction error handler threw an exception. '
                     + 'This places the crawler and its underlying storages into an unknown state and crawling will be terminated. '
@@ -544,6 +554,23 @@ export class BasicCrawler {
             }
         } finally {
             this.crawlingContexts.delete(crawlingContext.id);
+        }
+    }
+
+    /**
+     * Run async callback with given timeout and retry.
+     * @ignore
+     */
+    async _timeoutAndRetry(handler, timeout, error, maxRetries = 3, retried = 1) {
+        try {
+            await addTimeoutToPromise(handler, timeout, error);
+        } catch (e) {
+            if (retried <= maxRetries) { // we retry on any error, not just timeout
+                this.log.warning(`${e.message} (retrying ${retried}/${maxRetries})`);
+                return this._timeoutAndRetry(handler, timeout, error, maxRetries, retried + 1);
+            }
+
+            throw e;
         }
     }
 
@@ -662,7 +689,7 @@ export class BasicCrawler {
     }
 
     /**
-     * @param {Array<Hook>} hooks
+     * @param {Array<any>} hooks
      * @param  {*} args
      * @ignore
      * @protected