+++ src/crawlers/cheerio_crawler.js
@@ -1,6 +1,7 @@
 /* eslint-disable class-methods-use-this */
 import { readStreamToString, concatStreamToBuffer } from '@apify/utilities';
-import cheerio from 'cheerio';
+import { addTimeoutToPromise, tryCancel } from '@apify/timeout';
+import cheerio from 'cheerio'; // eslint-disable-line import/no-duplicates
 import contentTypeParser from 'content-type';
 import { DomHandler } from 'htmlparser2';
 import { WritableStream } from 'htmlparser2/lib/WritableStream';
@@ -9,8 +10,8 @@ import ow from 'ow';
 import util from 'util';
 import { TimeoutError } from 'got-scraping';
 import { BASIC_CRAWLER_TIMEOUT_BUFFER_SECS } from '../constants';
-import { addTimeoutToPromise, parseContentTypeFromResponse } from '../utils';
-import { requestAsBrowser } from '../utils_request';
+import { parseContentTypeFromResponse } from '../utils';
+import { requestAsBrowser } from '../utils_request'; // eslint-disable-line import/no-duplicates
 import { diffCookies, mergeCookies } from './crawler_utils';
 import { BasicCrawler } from './basic_crawler'; // eslint-disable-line import/no-duplicates
 import CrawlerExtension from './crawler_extension';
@@ -28,6 +29,9 @@ import { RequestQueue } from '../storages/request_queue';
 import { Session } from '../session_pool/session';
 import { SessionPoolOptions } from '../session_pool/session_pool';
 import { validators } from '../validators';
+import { RequestAsBrowserOptions } from '../utils_request';
+import { CheerioAPI } from 'cheerio';
+import { Hook } from './browser_crawler';
 /* eslint-enable no-unused-vars,import/named,import/no-duplicates,import/order */
 
 /**
@@ -470,12 +474,12 @@ class CheerioCrawler extends BasicCrawler {
         this.postResponseFunction = postResponseFunction;
         this.proxyConfiguration = proxyConfiguration;
         /**
-         * @type {Array<Hook>}
+         * @type {Array<any>}
          * @ignore
          * */
         this.preNavigationHooks = preNavigationHooks;
         /**
-         * @type {Array<Hook>}
+         * @type {Array<any>}
          * @ignore
          * */
         this.postNavigationHooks = [
@@ -543,8 +547,10 @@ class CheerioCrawler extends BasicCrawler {
         }
 
         await this._handleNavigation(crawlingContext);
+        tryCancel();
 
         const { dom, isXml, body, contentType, response } = await this._parseResponse(request, crawlingContext.response);
+        tryCancel();
 
         if (this.useSessionPool) {
             this._throwOnBlockedRequest(session, response.statusCode);
@@ -589,7 +595,7 @@ class CheerioCrawler extends BasicCrawler {
         });
 
         return addTimeoutToPromise(
-            this.userProvidedHandler(crawlingContext),
+            () => this.userProvidedHandler(crawlingContext),
             this.handlePageTimeoutMillis,
             `handlePageFunction timed out after ${this.handlePageTimeoutMillis / 1000} seconds.`,
         );
@@ -605,6 +611,7 @@ class CheerioCrawler extends BasicCrawler {
         if (this.prepareRequestFunction) {
             this.log.deprecated('Option "prepareRequestFunction" is deprecated. Use "preNavigationHooks" instead.');
             await this.prepareRequestFunction(crawlingContext);
+            tryCancel();
         }
 
         const requestAsBrowserOptions = {};
@@ -616,20 +623,24 @@ class CheerioCrawler extends BasicCrawler {
         const { request, session } = crawlingContext;
         const cookieSnapshot = request.headers.Cookie ?? request.headers.cookie;
         await this._executeHooks(this.preNavigationHooks, crawlingContext, requestAsBrowserOptions);
+        tryCancel();
         const proxyUrl = crawlingContext.proxyInfo && crawlingContext.proxyInfo.url;
         this._mergeRequestCookieDiff(request, cookieSnapshot, requestAsBrowserOptions);
 
         crawlingContext.response = await addTimeoutToPromise(
-            this._requestFunction({ request, session, proxyUrl, requestAsBrowserOptions }),
+            () => this._requestFunction({ request, session, proxyUrl, requestAsBrowserOptions }),
             this.requestTimeoutMillis,
             `request timed out after ${this.requestTimeoutMillis / 1000} seconds.`,
         );
+        tryCancel();
 
         await this._executeHooks(this.postNavigationHooks, crawlingContext, requestAsBrowserOptions);
+        tryCancel();
 
         if (this.postResponseFunction) {
             this.log.deprecated('Option "postResponseFunction" is deprecated. Use "postNavigationHooks" instead.');
             await this.postResponseFunction(crawlingContext);
+            tryCancel();
         }
     }
 
@@ -652,6 +663,7 @@ class CheerioCrawler extends BasicCrawler {
         const cookieDiff = diffCookies(request.url, cookieSnapshot, request.headers.Cookie ?? request.headers.cookie);
 
         if (cookieDiff.length > 0) {
+            requestAsBrowserOptions.headers ??= {};
             requestAsBrowserOptions.headers.Cookie = mergeCookies(request.url, [
                 requestAsBrowserOptions.headers.Cookie,
                 cookieDiff,
@@ -964,7 +976,7 @@ export default CheerioCrawler;
 
 /**
  * @typedef CheerioHandlePageInputs
- * @property {cheerio.Root} $
+ * @property {CheerioAPI} $
  *  The [Cheerio](https://cheerio.js.org/) object with parsed HTML.
  * @property {(string|Buffer)} body
  *  The request body of the web page.