+++ src/crawlers/puppeteer_crawler.js
@@ -10,11 +10,12 @@ import BrowserCrawler, { BrowserCrawlingContext } from './browser_crawler';
 import { ProxyConfiguration } from '../proxy_configuration';
 import { SessionPoolOptions } from '../session_pool/session_pool';
 import { CrawlingContext } from './basic_crawler';
-import { Page, DirectNavigationOptions } from 'puppeteer';
+import { Page } from 'puppeteer';
 import { RequestList } from '../request_list';
 import { RequestQueue } from '../storages/request_queue';
 import { AutoscaledPoolOptions } from '../autoscaling/autoscaled_pool';
 import { BrowserPoolOptions } from 'browser-pool';
+import { DirectNavigationOptions } from '../typedefs';
 /* eslint-enable no-unused-vars,import/named,import/no-duplicates,import/order */
 
 /**