+++ src/typedefs.js
@@ -19,7 +19,7 @@
  *   Time when the actor run finished. Contains `null` for running actors.
  * @property {string} status
  *   Status of the run. For possible values, see
- *   [Run lifecycle](https://docs.apify.com/actor/run#lifecycle)
+ *   [Run lifecycle](https://docs.apify.com/actors/running#lifecycle)
  *   in Apify actor documentation.
  * @property {Object<string, string>} meta
  *   Actor run meta-data. For example:
@@ -82,6 +82,22 @@
  *     }
  *   }
  *   ```
+ */
+
+/**
+ * @typedef {object} DirectNavigationOptions
+ * @property {number} [timeout]
+ *   Maximum operation time in milliseconds, defaults to 30 seconds, pass `0` to disable timeout. The
+ *   default value can be changed by using the browserContext.setDefaultNavigationTimeout(timeout),
+ *   browserContext.setDefaultTimeout(timeout), page.setDefaultNavigationTimeout(timeout) or
+ *   page.setDefaultTimeout(timeout) methods.
+ * @property {("domcontentloaded"|"load"|"networkidle")} [waitUntil]
+ *   When to consider operation succeeded, defaults to `load`. Events can be either:
+ *     - `'domcontentloaded'` - consider operation to be finished when the `DOMContentLoaded` event is fired.
+ *     - `'load'` - consider operation to be finished when the `load` event is fired.
+ *     - `'networkidle'` - consider operation to be finished when there are no network connections for at least `500` ms.
+ * @property {string} [referer]
+ *   Referer header value. If provided it will take preference over the referer header value set by page.setExtraHTTPHeaders(headers).
  */
 
 export {};