+++ test/browser_launchers/puppeteer_launcher.test.js
@@ -29,6 +29,11 @@ beforeAll(async () => {
         res.send(JSON.stringify(req.rawHeaders));
     });
 
+    app.get('/example', (req, res) => {
+        res.send(`<html><head><title>Example Domain</title></head></html>`);
+        res.status(200);
+    });
+
     app.all('/foo', (req, res) => {
         res.json({
             headers: req.headers,
@@ -207,16 +212,17 @@ describe('Apify.launchPuppeteer()', () => {
         let browser;
         const opts = {
             useChrome: true,
-            launchOptions: { headless: true },
+            launchOptions: { headless: true, timeout: 60e3 },
         };
 
         try {
             browser = await Apify.launchPuppeteer(opts);
             const page = await browser.newPage();
+            await page.setDefaultNavigationTimeout(0);
 
             // Add a test to go to an actual domain because we've seen issues
             // where pages would not load at all with Chrome.
-            await page.goto('https://example.com');
+            await page.goto(`http://${HOSTNAME}:${port}/example`);
             const title = await page.title();
             const version = await browser.version();