+++ test/puppeteer_utils.test.js
@@ -243,6 +243,7 @@ describe('Apify.utils.puppeteer', () => {
             const getResourcesLoadedFromWiki = async () => {
                 let downloadedBytes = 0;
                 const page = await browser.newPage();
+                await page.setDefaultNavigationTimeout(0);
                 // Cache all javascript files, png files and svg files
                 await Apify.utils.puppeteer.cacheResponses(page, cache, ['.js', /.+\.png/i, /.+\.svg/i]);
                 page.on('response', async (response) => {