+++ test/utils.test.js
@@ -7,6 +7,8 @@ import os from 'os';
 import cheerio from 'cheerio';
 import semver from 'semver';
 import { ENV_VARS } from '@apify/consts';
+import { addTimeoutToPromise } from '@apify/timeout';
+import { Log, LoggerText } from '@apify/log';
 import Apify from '../build/index';
 import * as utils from '../build/utils';
 import log from '../build/utils_log';
@@ -31,6 +33,13 @@ describe('utils.newClient()', () => {
 
         expect(client.token).toBe('token');
         expect(client.baseUrl).toBe('https://api.apify.com/v2');
+    });
+});
+
+describe('log export exposes custom loggers', () => {
+    test('works with log.LoggerText (#1238)', () => {
+        expect(Apify.utils.log).toBeInstanceOf(Log);
+        expect(Apify.utils.log.LoggerText).toBe(LoggerText);
     });
 });
 
@@ -812,8 +821,8 @@ describe('utils.addTimeoutToPromise()', () => {
     test('should timeout', async () => {
         const clock = sinon.useFakeTimers();
         try {
-            const p = utils.addTimeoutToPromise(
-                new Promise((r) => setTimeout(r, 500)),
+            const p = addTimeoutToPromise(
+                () => new Promise((r) => setTimeout(r, 500)),
                 100,
                 'Timed out.',
             );
@@ -830,8 +839,8 @@ describe('utils.addTimeoutToPromise()', () => {
     test('should not timeout too soon', async () => {
         const clock = sinon.useFakeTimers();
         try {
-            const p = utils.addTimeoutToPromise(
-                new Promise((r) => setTimeout(() => r('Done'), 100)),
+            const p = addTimeoutToPromise(
+                () => new Promise((r) => setTimeout(() => r('Done'), 100)),
                 500,
                 'Timed out.',
             );