+++ website/siteConfig.js
@@ -8,17 +8,6 @@
 // See https://docusaurus.io/docs/site-config for all the possible
 // site configuration options.
 
-// List of projects/orgs using your project for the users page.
-// const users = [
-//     {
-//         caption: 'User1',
-//         // You will need to prepend the image path with your baseUrl
-//         // if it is not '/', like: '/test-site/img/docusaurus.svg'.
-//         image: '/img/docusaurus.svg',
-//         infoLink: 'https://www.facebook.com',
-//         pinned: true,
-//     },
-// ];
 const extlink = require('remarkable-extlink');
 
 const repoUrl = 'https://github.com/apify/apify-js';
@@ -32,16 +21,10 @@ const siteConfig = {
     url: 'https://sdk.apify.com', // Your website URL
     cname: 'sdk.apify.com',
     baseUrl: '/', // Base URL for your project */
-    // For github.io type URLs, you would set the url and baseUrl like:
-    //   url: 'https://facebook.github.io',
-    //   baseUrl: '/test-site/',
 
     // Used for publishing and more
     projectName: 'apify-js',
     organizationName: 'apify',
-    // For top-level user or org sites, the organization is still the same.
-    // e.g., for the https://JoelMarcey.github.io site, it would be set like...
-    //   organizationName: 'JoelMarcey'
 
     // For no header links in the top nav bar -> headerLinks: [],
     headerLinks: [
@@ -53,9 +36,6 @@ const siteConfig = {
         // { page: 'help', label: 'Help' },
         // { blog: true, label: 'Blog' },
     ],
-
-    // If you have users set above, you add it here:
-    // users,
 
     /* path to images for header/footer */
     headerIcon: 'img/apify_logo.svg',
@@ -77,20 +57,6 @@ const siteConfig = {
             facetFilters: ['version:VERSION'],
         },
     },
-
-    /* Custom fonts for website */
-    /*
-  fonts: {
-    myFont: [
-      "Times New Roman",
-      "Serif"
-    ],
-    myOtherFont: [
-      "-apple-system",
-      "system-ui"
-    ]
-  },
-  */
 
     // This copyright info is used in /core/Footer.js and blog RSS/Atom feeds.
     copyright: `Copyright © ${new Date().getFullYear()} Apify Technologies s.r.o.`,
@@ -135,6 +101,7 @@ const siteConfig = {
         "1.0.0": [
             "api/puppeteer-pool",
             "typedefs/puppeteer-pool-options",
+            "typedefs/launch-puppeteer",
             "typedefs/launch-puppeteer-function",
             "typedefs/launch-puppeteer-options",
             "typedefs/puppeteer-goto",
@@ -146,9 +113,6 @@ const siteConfig = {
     ogImage: 'img/apify_og_SDK.png',
     twitterImage: 'img/apify_og_SDK.png',
 
-    // You may provide arbitrary config keys to be used as needed by your
-    // template. For example, if you need your repo's URL...
-    //   repoUrl: 'https://github.com/facebook/test-site',
     repoUrl,
 };